+++ src/app/components/resetPassword/resetPasswordService.js
@@ -1,7 +1,7 @@
 'use strict';
 
 angular.module('liveopsConfigPanel')
-  .service('ResetPassword', ['$http', 'apiHostname', function($http, apiHostname) {
+  .service('ResetPassword', ['$http', 'apiHostname', '$location', function($http, apiHostname, $location) {
     var service = {};
 
     service.initiateReset = function(userId) {
@@ -17,6 +17,9 @@ angular.module('liveopsConfigPanel')
     service.setNewPassword = function(userId, newPass) {
       return $http({
         method: 'PUT',
+        headers: {
+          'Authorization': 'Token ' + $location.search().token
+        },
         url: apiHostname + '/v1/users/' + userId,
         data: {
           'password': newPass